Separated log functions and Liberty-to-JSON functions from quicklogic_timings_importer
